chore: Move Mods tab out of LuaDebugger
TODO: Decide if the Lua file operations(Open, New buttons) should still exist, and if they should, they should redirect automatically to the Lua Debugger tab since that's where the editor is.

TODO: Add BP mods.
