+++ UE4SS/include/GUI/LuaDebugger.hpp
@@ -289,7 +289,6 @@ namespace RC::GUI
         auto render_controls() -> void;
         auto render_debug_view() -> void;
         auto render_script_editor() -> void;
-        auto render_mods_tab() -> void;
         auto render_breakpoints_panel() -> void;
         auto render_repl() -> void;
         auto render_value_tree(std::vector<LuaValueNode>& nodes) -> void;
@@ -321,44 +320,7 @@ namespace RC::GUI
         // Helper to get table entries at a given path
         static auto get_table_entries_at_path(lua_State* L, const std::string& path) -> std::vector<std::pair<std::string, LuaStackSlot>>;
 
-        // TODO: Move to Mod.hpp
-        enum class ModType
-        {
-            Invalid,
-            LuaMod,
-            CppMod,
-        };
-
-        // Mod management
-        struct ModInfo
-        {
-            std::string name;
-            std::filesystem::path path;
-            bool enabled_via_txt{false};
-            bool enabled_via_mods_txt{false};
-            bool is_running{false};
-            ModType mod_type{ModType::Invalid};
-
-            bool is_enabled() const { return enabled_via_txt || enabled_via_mods_txt; }
-        };
-
-        std::vector<ModInfo> m_discovered_mods;
-        bool m_mods_list_dirty{true};
-        std::string m_new_mod_name;
-        bool m_show_create_mod_popup{false};
-        std::string m_new_file_name;
-        bool m_show_create_file_popup{false};
-        bool m_add_require_to_main{true};
-        std::filesystem::path m_create_file_mod_path;
         int m_pending_editor_tab_switch{-1};
-
-        auto refresh_mods_list() -> void;
-        auto create_new_mod(const std::string& name) -> bool;
-        auto create_new_file(const std::string& mod_path, const std::string& filename, bool add_require_to_main) -> bool;
-        auto set_mod_enabled(const std::filesystem::path& mod_path, bool enabled) -> void;
-        auto restart_mod_by_name(const std::string& mod_name, ModType mod_type) -> void;
-        auto uninstall_mod_by_name(const std::string& mod_name, ModType mod_type) -> void;
-        auto start_mod_by_path(const std::filesystem::path& mod_path, ModType mod_type) -> void;
     };
 
 } // namespace RC::GUI
