+++ UE4SS/src/GUI/LuaDebugger.cpp
@@ -1706,12 +1706,6 @@ namespace RC::GUI
                 ImGui::EndTabItem();
             }
 
-            if (ImGui::BeginTabItem(ICON_FA_CUBES " Mods"))
-            {
-                render_mods_tab();
-                ImGui::EndTabItem();
-            }
-
             ImGui::EndTabBar();
         }
 
@@ -3372,7 +3366,7 @@ namespace RC::GUI
             {
                 if (ImGui::Button(ICON_FA_SYNC " Restart"))
                 {
-                    restart_mod_by_name(mod_name, ModType::LuaMod);
+                    UE4SSProgram::get_program().queue_reinstall_lua_mod_by_name(mod_name);
                 }
                 if (ImGui::IsItemHovered())
                 {
@@ -3383,7 +3377,7 @@ namespace RC::GUI
             {
                 if (ImGui::Button(ICON_FA_PLAY " Start"))
                 {
-                    start_mod_by_path(mod_path, ModType::LuaMod);
+                    UE4SSProgram::get_program().queue_start_lua_mod_by_path(mod_path);
                 }
                 if (ImGui::IsItemHovered())
                 {
@@ -3609,666 +3603,4 @@ namespace RC::GUI
             }
         }
     }
-
-    auto LuaDebugger::refresh_mods_list() -> void
-    {
-        m_discovered_mods.clear();
-
-        auto& program = UE4SSProgram::get_program();
-        auto& mods_dirs = program.get_mods_directories();
-        auto mods_txt_entries = program.get_mods_txt_entries();
-
-        std::set<std::string> running_mod_names;
-        for (const auto& mod : program.m_mods)
-        {
-            if (mod->is_started())
-            {
-                auto running_identifier = to_string(mod->get_name());
-                if (dynamic_cast<LuaMod*>(mod.get()))
-                {
-                    running_identifier += "_LuaMod";
-                }
-                else
-                {
-                    running_identifier += "_CppMod";
-                }
-                running_mod_names.insert(running_identifier);
-            }
-        }
-
-        std::set<std::string> seen_mods;
-
-        for (const auto& mods_dir : mods_dirs)
-        {
-            if (!std::filesystem::exists(mods_dir))
-            {
-                continue;
-            }
-
-            std::error_code ec;
-            for (const auto& entry : std::filesystem::directory_iterator(mods_dir, ec))
-            {
-                if (!entry.is_directory(ec) || ec)
-                {
-                    continue;
-                }
-
-                std::string mod_name = entry.path().stem().string();
-
-                if (seen_mods.count(mod_name))
-                {
-                    continue;
-                }
-                seen_mods.insert(mod_name);
-
-                ModInfo info;
-                info.name = mod_name;
-                info.path = entry.path();
-                info.enabled_via_txt = std::filesystem::exists(entry.path() / "enabled.txt", ec);
-
-                const bool has_main_lua = std::filesystem::exists(entry.path() / "Scripts" / "main.lua", ec);
-                const bool has_main_dll = std::filesystem::exists(entry.path() / "dlls" / "main.dll", ec) ||
-                    std::filesystem::exists(entry.path() / "dlls" / fmt::format("{}.dll", mod_name), ec);
-
-                if (has_main_lua || has_main_dll)
-                {
-                    auto it = mods_txt_entries.find(mod_name);
-                    if (it != mods_txt_entries.end())
-                    {
-                        info.enabled_via_mods_txt = it->second;
-                    }
-                }
-
-                if (has_main_lua)
-                {
-                    info.mod_type = ModType::LuaMod;
-                    info.is_running = running_mod_names.contains(mod_name + "_LuaMod") > 0;
-                    m_discovered_mods.push_back(info);
-                }
-
-                if (has_main_dll)
-                {
-                    info.mod_type = ModType::CppMod;
-                    info.is_running = running_mod_names.contains(mod_name + "_CppMod") > 0;
-                    m_discovered_mods.push_back(info);
-                }
-            }
-        }
-
-        std::sort(m_discovered_mods.begin(),
-                  m_discovered_mods.end(),
-                  [](const auto& a, const auto& b) {
-                      return a.name < b.name;
-                  });
-
-        m_mods_list_dirty = false;
-    }
-
-    auto LuaDebugger::set_mod_enabled(const std::filesystem::path& mod_path, bool enabled) -> void
-    {
-        std::filesystem::path enabled_file = mod_path / "enabled.txt";
-        std::error_code ec;
-
-        if (enabled)
-        {
-            std::ofstream file(enabled_file);
-            file.close();
-        }
-        else
-        {
-            std::filesystem::remove(enabled_file, ec);
-        }
-
-        m_mods_list_dirty = true;
-    }
-
-    auto LuaDebugger::restart_mod_by_name(const std::string& mod_name, ModType mod_type) -> void
-    {
-        if (mod_type == ModType::LuaMod)
-        {
-            UE4SSProgram::get_program().queue_reinstall_lua_mod_by_name(mod_name);
-        }
-        else
-        {
-            UE4SSProgram::get_program().queue_reinstall_cpp_mod_by_name(mod_name);
-        }
-        if (LuaDebugger::has_instance())
-        {
-            LuaDebugger::get().m_mods_list_dirty = true;
-        }
-    }
-
-    auto LuaDebugger::uninstall_mod_by_name(const std::string& mod_name, ModType mod_type) -> void
-    {
-        if (mod_type == ModType::LuaMod)
-        {
-            UE4SSProgram::get_program().queue_uninstall_lua_mod_by_name(mod_name);
-        }
-        else
-        {
-            UE4SSProgram::get_program().queue_uninstall_cpp_mod_by_name(mod_name);
-        }
-        if (LuaDebugger::has_instance())
-        {
-            LuaDebugger::get().m_mods_list_dirty = true;
-        }
-    }
-
-    auto LuaDebugger::start_mod_by_path(const std::filesystem::path& mod_path, ModType mod_type) -> void
-    {
-        if (mod_type == ModType::LuaMod)
-        {
-            UE4SSProgram::get_program().queue_start_lua_mod_by_path(mod_path);
-        }
-        else
-        {
-            UE4SSProgram::get_program().queue_start_cpp_mod_by_path(mod_path);
-        }
-        if (LuaDebugger::has_instance())
-        {
-            LuaDebugger::get().m_mods_list_dirty = true;
-        }
-    }
-
-    auto LuaDebugger::create_new_mod(const std::string& name) -> bool
-    {
-        if (name.empty())
-        {
-            return false;
-        }
-
-        auto& program = UE4SSProgram::get_program();
-        auto& mods_dirs = program.get_mods_directories();
-
-        if (mods_dirs.empty())
-        {
-            return false;
-        }
-
-        std::filesystem::path mod_path = mods_dirs[0] / name;
-        std::filesystem::path scripts_path = mod_path / "Scripts";
-        std::filesystem::path main_lua_path = scripts_path / "main.lua";
-
-        std::error_code ec;
-
-        if (std::filesystem::exists(mod_path, ec))
-        {
-            Output::send<LogLevel::Warning>(STR("Mod '{}' already exists\n"), to_generic_string(name));
-            return false;
-        }
-
-        std::filesystem::create_directories(scripts_path, ec);
-        if (ec)
-        {
-            Output::send<LogLevel::Error>(STR("Failed to create mod directory: {}\n"), to_generic_string(ec.message()));
-            return false;
-        }
-
-        std::ofstream main_file(main_lua_path);
-        if (!main_file.is_open())
-        {
-            Output::send<LogLevel::Error>(STR("Failed to create main.lua\n"));
-            return false;
-        }
-
-        main_file << "-- " << name << " mod\n";
-        main_file << "-- Created by UE4SS Lua Debugger\n\n";
-        main_file << "print(\"[" << name << "] Mod loaded!\")\n";
-        main_file.close();
-
-        std::ofstream enabled_file(mod_path / "enabled.txt");
-        enabled_file.close();
-
-        m_mods_list_dirty = true;
-
-        // New mod is not running, clear selected state so dropdown uses filesystem scan
-        m_selected_state = nullptr;
-
-        m_script_edit_path = main_lua_path.string();
-        const LuaScriptFile* script = load_script(m_script_edit_path);
-        if (script && script->loaded)
-        {
-            m_script_editor.SetText(script->content);
-            m_script_original_content = m_script_editor.GetText();
-            m_script_is_dirty = false;
-        }
-
-        m_pending_editor_tab_switch = 1;
-
-        Output::send(STR("Created new mod '{}'\n"), to_generic_string(name));
-        return true;
-    }
-
-    auto LuaDebugger::create_new_file(const std::string& mod_path, const std::string& filename, bool add_require_to_main) -> bool
-    {
-        if (filename.empty() || mod_path.empty())
-        {
-            return false;
-        }
-
-        std::filesystem::path scripts_path = std::filesystem::path(mod_path) / "Scripts";
-        std::filesystem::path file_path = scripts_path / filename;
-
-        if (!file_path.has_extension())
-        {
-            file_path += ".lua";
-        }
-
-        std::error_code ec;
-        if (std::filesystem::exists(file_path, ec))
-        {
-            Output::send<LogLevel::Warning>(STR("File '{}' already exists\n"), to_generic_string(filename));
-            return false;
-        }
-
-        std::ofstream file(file_path);
-        if (!file.is_open())
-        {
-            Output::send<LogLevel::Error>(STR("Failed to create file '{}'\n"), to_generic_string(filename));
-            return false;
-        }
-
-        std::string module_name = file_path.stem().string();
-        file << "-- " << module_name << " module\n\n";
-        file << "local M = {}\n\n";
-        file << "return M\n";
-        file.close();
-
-        // Add require() to main.lua if requested
-        if (add_require_to_main)
-        {
-            std::filesystem::path main_lua_path = scripts_path / "main.lua";
-            if (std::filesystem::exists(main_lua_path))
-            {
-                std::ifstream main_file_read(main_lua_path);
-                std::string main_content((std::istreambuf_iterator<char>(main_file_read)), std::istreambuf_iterator<char>());
-                main_file_read.close();
-
-                std::string require_line = "local " + module_name + " = require(\"" + module_name + "\")\n";
-
-                // Check if require already exists
-                if (main_content.find("require(\"" + module_name + "\")") == std::string::npos)
-                {
-                    // Find the best place to insert - after existing requires or at the top
-                    size_t insert_pos = 0;
-                    size_t last_require_pos = main_content.rfind("require(");
-                    if (last_require_pos != std::string::npos)
-                    {
-                        // Find end of line after last require
-                        size_t newline_pos = main_content.find('\n', last_require_pos);
-                        if (newline_pos != std::string::npos)
-                        {
-                            insert_pos = newline_pos + 1;
-                        }
-                    }
-                    else
-                    {
-                        // No existing requires, find end of any initial comments
-                        size_t pos = 0;
-                        while (pos < main_content.size())
-                        {
-                            // Skip whitespace
-                            while (pos < main_content.size() && (main_content[pos] == ' ' || main_content[pos] == '\t' || main_content[pos] == '\n' || main_content[pos] == '\r'))
-                                pos++;
-
-                            if (pos < main_content.size() && main_content[pos] == '-' && pos + 1 < main_content.size() && main_content[pos + 1] == '-')
-                            {
-                                // Skip comment line
-                                size_t newline = main_content.find('\n', pos);
-                                if (newline != std::string::npos)
-                                    pos = newline + 1;
-                                else
-                                    break;
-                            }
-                            else
-                            {
-                                break;
-                            }
-                        }
-                        insert_pos = pos;
-                    }
-
-                    main_content.insert(insert_pos, require_line);
-
-                    std::ofstream main_file_write(main_lua_path);
-                    main_file_write << main_content;
-                    main_file_write.close();
-
-                    // Clear cache for main.lua so it reloads
-                    {
-                        std::lock_guard<std::mutex> lock(m_scripts_mutex);
-                        m_script_cache.erase(main_lua_path.string());
-                    }
-
-                    Output::send(STR("Added require() for '{}' to main.lua\n"), to_generic_string(module_name));
-                }
-            }
-        }
-
-        {
-            std::lock_guard<std::mutex> lock(m_scripts_mutex);
-            m_script_cache.erase(file_path.string());
-        }
-
-        // Check if mod is running; if not, clear selected state so dropdown uses filesystem scan
-        bool mod_is_running = false;
-        std::string mod_name = std::filesystem::path(mod_path).stem().string();
-        for (const auto& mod : UE4SSProgram::get_program().m_mods)
-        {
-            auto* lua_mod = dynamic_cast<LuaMod*>(mod.get());
-            if (lua_mod && to_string(lua_mod->get_name()) == mod_name && lua_mod->is_started())
-            {
-                mod_is_running = true;
-                break;
-            }
-        }
-        if (!mod_is_running)
-        {
-            m_selected_state = nullptr;
-        }
-
-        m_script_edit_path = file_path.string();
-        const LuaScriptFile* script = load_script(m_script_edit_path);
-        if (script && script->loaded)
-        {
-            m_script_editor.SetText(script->content);
-            m_script_original_content = m_script_editor.GetText();
-            m_script_is_dirty = false;
-        }
-
-        m_pending_editor_tab_switch = 1;
-
-        Output::send(STR("Created new file '{}'\n"), to_generic_string(file_path.string()));
-        return true;
-    }
-
-    auto LuaDebugger::render_mods_tab() -> void
-    {
-        if (m_mods_list_dirty)
-        {
-            refresh_mods_list();
-        }
-
-        if (ImGui::Button(ICON_FA_SYNC " Refresh"))
-        {
-            m_mods_list_dirty = true;
-        }
-
-        ImGui::SameLine();
-
-        if (ImGui::Button(ICON_FA_PLUS " New Mod"))
-        {
-            m_show_create_mod_popup = true;
-            m_new_mod_name.clear();
-        }
-
-        ImGui::SameLine();
-
-        auto& gui = UE4SSProgram::get_program().get_debugging_ui();
-        bool event_busy = gui.m_event_thread_busy || !UE4SSProgram::get_program().can_process_events();
-        if (event_busy)
-        {
-            ImGui::BeginDisabled(true);
-        }
-        if (ImGui::Button(ICON_FA_REDO " Restart All Mods"))
-        {
-            gui.m_event_thread_busy = true;
-            UE4SSProgram::get_program().queue_event([&gui, this]() {
-                UE4SSProgram::get_program().queue_reinstall_mods();
-                gui.m_event_thread_busy = false;
-                m_mods_list_dirty = true;
-            });
-        }
-        if (event_busy)
-        {
-            ImGui::EndDisabled();
-        }
-
-        ImGui::SameLine();
-        ImGui::TextDisabled("(%zu mods)", m_discovered_mods.size());
-
-        ImGui::Separator();
-
-        ImGui::BeginChild("ModsList", ImVec2(0, 0), true, ImGuiWindowFlags_HorizontalScrollbar);
-
-        for (size_t i = 0; i < m_discovered_mods.size(); ++i)
-        {
-            auto& mod = m_discovered_mods[i];
-
-            ImGui::PushID(static_cast<int>(i));
-
-            bool enabled = mod.is_enabled();
-            if (ImGui::Checkbox("##enabled", &enabled))
-            {
-                set_mod_enabled(mod.path, enabled);
-                mod.enabled_via_txt = enabled;
-            }
-
-            if (ImGui::IsItemHovered())
-            {
-                std::string tooltip = "Enable/disable on next reload";
-                if (mod.enabled_via_mods_txt && !mod.enabled_via_txt)
-                {
-                    tooltip += "\n(Currently enabled via mods.txt)";
-                }
-                ImGui::SetTooltip("%s", tooltip.c_str());
-            }
-
-            ImGui::SameLine();
-
-            if (mod.is_running)
-            {
-                ImGui::TextColored(ImVec4(0.4f, 1.0f, 0.4f, 1.0f), ICON_FA_PLAY);
-                ImGui::SameLine();
-            }
-
-            const auto mod_name = fmt::format("{} ({})", mod.name, mod.mod_type == ModType::LuaMod ? "Lua" : "Cpp");
-            if (mod.enabled_via_mods_txt && !mod.enabled_via_txt)
-            {
-                ImGui::TextColored(ImVec4(0.7f, 0.7f, 1.0f, 1.0f), "%s", mod_name.c_str());
-            }
-            else
-            {
-                ImGui::Text("%s", mod_name.c_str());
-            }
-
-            // Use consistent button width for alignment (widest case: Open + New + Restart + Uninstall)
-            float button_width = scaled(320.0f);
-            ImGui::SameLine(ImGui::GetContentRegionAvail().x - button_width);
-
-            if (mod.mod_type != ModType::LuaMod)
-            {
-                ImGui::BeginDisabled();
-            }
-            if (ImGui::SmallButton(ICON_FA_FOLDER_OPEN " Open"))
-            {
-                std::filesystem::path scripts_path = mod.path / "Scripts";
-                std::filesystem::path main_lua = scripts_path / "main.lua";
-                if (std::filesystem::exists(main_lua))
-                {
-                    // Clear selected state if mod isn't running so dropdown uses filesystem scan
-                    if (!mod.is_running)
-                    {
-                        m_selected_state = nullptr;
-                    }
-                    m_script_edit_path = main_lua.string();
-                    const LuaScriptFile* script = load_script(m_script_edit_path);
-                    if (script && script->loaded)
-                    {
-                        m_script_editor.SetText(script->content);
-                        m_script_original_content = m_script_editor.GetText();
-                        m_script_is_dirty = false;
-                        m_pending_editor_tab_switch = 1;
-                    }
-                }
-            }
-            if (mod.mod_type != ModType::LuaMod)
-            {
-                ImGui::EndDisabled();
-            }
-
-            ImGui::SameLine();
-
-            if (mod.mod_type != ModType::LuaMod)
-            {
-                ImGui::BeginDisabled();
-            }
-            if (ImGui::SmallButton(ICON_FA_FILE_ALT " New"))
-            {
-                m_show_create_file_popup = true;
-                m_new_file_name.clear();
-                m_add_require_to_main = true;
-                m_create_file_mod_path = mod.path;
-            }
-            if (mod.mod_type != ModType::LuaMod)
-            {
-                ImGui::EndDisabled();
-            }
-
-            // Show different buttons based on whether mod is running
-            if (mod.is_running)
-            {
-                ImGui::SameLine();
-
-                if (event_busy)
-                {
-                    ImGui::BeginDisabled(true);
-                }
-
-                if (ImGui::SmallButton(ICON_FA_SYNC " Restart"))
-                {
-                    restart_mod_by_name(mod.name, mod.mod_type);
-                }
-                if (ImGui::IsItemHovered())
-                {
-                    ImGui::SetTooltip("Uninstall and reinstall this mod");
-                }
-
-                ImGui::SameLine();
-
-                if (ImGui::SmallButton(ICON_FA_STOP " Uninstall"))
-                {
-                    uninstall_mod_by_name(mod.name, mod.mod_type);
-                }
-                if (ImGui::IsItemHovered())
-                {
-                    ImGui::SetTooltip("Uninstall this mod (will not run until next full reload)");
-                }
-
-                if (event_busy)
-                {
-                    ImGui::EndDisabled();
-                }
-            }
-            else
-            {
-                // Show Start button for mods that are not running
-                ImGui::SameLine();
-
-                if (event_busy)
-                {
-                    ImGui::BeginDisabled(true);
-                }
-
-                if (ImGui::SmallButton(ICON_FA_PLAY " Start"))
-                {
-                    start_mod_by_path(mod.path, mod.mod_type);
-                }
-                if (ImGui::IsItemHovered())
-                {
-                    ImGui::SetTooltip("Start this mod without reloading all mods");
-                }
-
-                if (event_busy)
-                {
-                    ImGui::EndDisabled();
-                }
-            }
-
-            if (ImGui::IsItemHovered())
-            {
-                ImGui::BeginTooltip();
-                ImGui::Text("Path: %s", mod.path.string().c_str());
-                ImGui::EndTooltip();
-            }
-
-            ImGui::PopID();
-        }
-
-        ImGui::EndChild();
-
-        if (m_show_create_mod_popup)
-        {
-            ImGui::OpenPopup("Create New Mod");
-            m_show_create_mod_popup = false;
-        }
-
-        if (m_show_create_file_popup)
-        {
-            ImGui::OpenPopup("Create New File");
-            m_show_create_file_popup = false;
-        }
-
-        if (ImGui::BeginPopupModal("Create New Mod", nullptr, ImGuiWindowFlags_AlwaysAutoResize))
-        {
-            ImGui::Text("Enter name for new mod:");
-            ImGui::Separator();
-
-            ImGui::SetNextItemWidth(scaled(300.0f));
-            bool enter_pressed = ImGui::InputText("##newmodname", &m_new_mod_name, ImGuiInputTextFlags_EnterReturnsTrue);
-
-            if ((ImGui::Button("Create", ImVec2(scaled(120.0f), 0)) || enter_pressed) && !m_new_mod_name.empty())
-            {
-                if (create_new_mod(m_new_mod_name))
-                {
-                    ImGui::CloseCurrentPopup();
-                }
-            }
-
-            ImGui::SameLine();
-
-            if (ImGui::Button("Cancel", ImVec2(scaled(120.0f), 0)))
-            {
-                ImGui::CloseCurrentPopup();
-            }
-
-            ImGui::EndPopup();
-        }
-
-        if (ImGui::BeginPopupModal("Create New File", nullptr, ImGuiWindowFlags_AlwaysAutoResize))
-        {
-            std::string mod_name = m_create_file_mod_path.stem().string();
-            ImGui::Text("Create new file in %s:", mod_name.c_str());
-            ImGui::Separator();
-
-            ImGui::SetNextItemWidth(scaled(300.0f));
-            bool enter_pressed = ImGui::InputText("##newfilename", &m_new_file_name, ImGuiInputTextFlags_EnterReturnsTrue);
-
-            if (!m_new_file_name.empty() && m_new_file_name.find(".lua") == std::string::npos)
-            {
-                ImGui::SameLine();
-                ImGui::TextDisabled(".lua");
-            }
-
-            ImGui::Checkbox("Add require() to main.lua", &m_add_require_to_main);
-
-            if ((ImGui::Button("Create", ImVec2(scaled(120.0f), 0)) || enter_pressed) && !m_new_file_name.empty())
-            {
-                if (create_new_file(m_create_file_mod_path.string(), m_new_file_name, m_add_require_to_main))
-                {
-                    ImGui::CloseCurrentPopup();
-                }
-            }
-
-            ImGui::SameLine();
-
-            if (ImGui::Button("Cancel", ImVec2(scaled(120.0f), 0)))
-            {
-                ImGui::CloseCurrentPopup();
-            }
-
-            ImGui::EndPopup();
-        }
-    }
-
 } // namespace RC::GUI
